requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
python-dotenv
pymysql
email-validator
